Implement PPI-based value search/replace
This adds a new option 'overwrite'.  When enabled, and a normal
'# VERSION' comment is found, this will check to see if the comment
was preceeded (on the same line) by tokens matching a version
assignment, and if so, overwrite the value token with the new version.
If the tokens do not match this pattern, it falls back to the normal
behavior.